Move SkillManager from __main__ to separate file
